fix(blobstore/nats): cap MemoryOutputPipe at u32::MAX for wasm32 compat
The wasi::io std::io::Write impl calls check_write() and converts its u64
return value via try_into::<usize>(). On a 32-bit wasm component, usize is
u32. When the pipe capacity is usize::MAX on a 64-bit host (2^64-1), the
try_into() call fails with 'out of range integral type conversion attempted',
causing every upload to return HTTP 500.

Cap the pipe capacity at u32::MAX (4 GiB) so the value roundtrips safely
through the WIT u64 interface to a 32-bit wasm usize.

+++ crates/wash-runtime/src/plugin/wasi_blobstore/nats.rs
@@ -645,7 +645,10 @@ impl<'a> bindings::wasi::blobstore::types::HostOutgoingValue for ActiveCtx<'a> {
     #[instrument(skip(self))]
     async fn new_outgoing_value(&mut self) -> anyhow::Result<Resource<OutgoingValueHandle>> {
         let handle = OutgoingValueHandle {
-            pipe: MemoryOutputPipe::new(usize::MAX),
+            // Cap at u32::MAX so the pipe capacity fits in a 32-bit wasm usize.
+            // The wasi::io std::io::Write impl converts check_write()'s u64 via
+            // try_into::<usize>(), which fails on wasm32 when the value exceeds u32::MAX.
+            pipe: MemoryOutputPipe::new(u32::MAX as usize),
             container: None,
             object_name: None,
         };
